Pharmaship 1.1

+++ purchase/views.py
@@ -141,7 +141,8 @@ def status(request, requisition_id):
             requisition.save()
             data = json.dumps({'success': _('Data updated'),
                                'id': requisition_id,
-                               'status': requisition.get_status_display()})
+                               'status': requisition.get_status_display(),
+                               'code': requisition.status})
             return HttpResponse(data, content_type="application/json")
         else:
             errors = dict(
@@ -183,6 +184,12 @@ def item_add(request, requisition_id):
     if request.method == 'POST':
         form = forms.AddItemForm(request.POST)
         if form.is_valid():
+            # Check if there is an existing item
+            if len(requisition.item_set.filter(object_id=form.cleaned_data['object_id'])) > 0:
+                # The item already exists
+                data = json.dumps({'error': _('This item is already in the requisition.')})
+                return HttpResponseBadRequest(data, content_type="application/json")
+
             # Create a new requisition item
             item = models.Item()
             item.content_type = ContentType.objects.get_for_id(requisition.item_type)
@@ -201,8 +208,8 @@ def item_add(request, requisition_id):
             errors = dict(
                 [(k, [unicode(e) for e in v]) for k, v in form.errors.items()]
             )
-            data = json.dumps({'error': _('Something went wrong!'),
-                               'details': errors})
+            # Return only the first error to be displayed inline
+            data = json.dumps({'error': errors[errors.keys()[0]]})
             return HttpResponseBadRequest(data,
                                           content_type="application/json")
 
@@ -214,6 +221,12 @@ def delete(request, requisition_id):
     requisition = get_object_or_404(models.Requisition, pk=requisition_id)
     # Parsing the form
     if request.method == 'POST' and request.is_ajax():
+        if requisition.status > 0:
+            # The requisition as already been sent
+            data = json.dumps({'error': _('Requisition already sent!')})
+            return HttpResponseBadRequest(data,
+                                          content_type="application/json")
+
         requisition.delete()
         data = json.dumps({'success': _('Data updated'),
                            'url': reverse('purchase:index')})
@@ -233,7 +246,7 @@ def delete(request, requisition_id):
                 _("You are going to delete the requisition:"),
                 requisition.name,
             ),
-        'foot_text': '<h4 class=\'text-danger\'>{0}</h4>'.format(
+        'foot_text': u'<h4 class=\'text-danger\'>{0}</h4>'.format(
             _('The deletion is permanent!')),
         'callback': 'redirect',
         },
@@ -317,6 +330,12 @@ def edit(request, requisition_id):
     requisition = get_object_or_404(models.Requisition, pk=requisition_id)
 
     if request.method == "POST" and request.is_ajax():
+        if requisition.status > 0:
+            # The requisition as already been sent
+            data = json.dumps({'error': _('Requisition already sent!')})
+            return HttpResponseBadRequest(data,
+                                          content_type="application/json")
+
         form = forms.RequisitionNameForm(request.POST)
         if form.is_valid():
             # Process the data in form.cleaned_data
